* ThreadSearch: don't prefix tilde to search word(#742, thanks Gokul Krishnan)
I'm suggesting a modification in the ThreadSearch "Find Occurrences"
behavior for words that are preceded by a tilde. In such cases,
ThreadSearch automatically prepends the tilde to the search keyword
(i.e., assuming it's a C++ destructor). But the tilde is also a bit-wise
negation operator.

Consider,

int var = 0;
var = ~var;

Upon right-clicking on 'var' (on the right-hand side), the context menu
shows Find occurrences of: '~var', which is wrong. So I think the
automatic prefixing should be disabled. The user can still select an
arbitrary range of text to specifically include such symbols.